requirements.txt
aioredis==1.3.1    
arrow==1.2.3    
asgiref==3.5.2    
async-timeout==4.0.2    
attrs==22.1.0   
autobahn==22.7.1   
Automat==22.10.0    
binaryornot==0.4.4    
certifi==2022.9.24
cffi==1.15.1
channels==3.0.4
channels-redis==3.3.1
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
constantly==15.1.0
cookiecutter==2.1.1
cryptography==38.0.3
daphne==3.0.2
Deprecated==1.2.13
distlib==0.3.6
Django==4.0
django_debug_toolbar==3.8.1
filelock==3.8.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
Jinja2==3.1.2
jinja2-time==0.2.0
MarkupSafe==2.1.1
msgpack==1.0.4
mysqlclient==2.1.1
packaging==21.3
Pillow==9.3.0
pip==22.2.2
platformdirs==2.5.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pyOpenSSL==22.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-slugify==6.1.2
pytz==2022.6
PyYAML==6.0
redis==4.3.4
requests==2.28.1
service-identity==21.1.0
setuptools==58.1.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.4.0
tzdata==2022.2
urllib3==1.26.12
virtualenv==20.16.6
wget==3.2
wrapt==1.14.1
zope.interface==5.5.1